refactor: remove subdirectory, sort imports, Dockerfile changes
