+++ packages/topo-imagery-common/src/aws/aws_credential_source.py
