+++ packages/topo-imagery-common/src/aws/tests/aws_helper_test.py
@@ -1,6 +1,5 @@
 from pytest_subtests import SubTests
-
-from topo_imagery_common.aws.aws_helper import parse_path
+from aws.aws_helper import parse_path
 
 
 def test_parse_path_s3(subtests: SubTests) -> None:
