+++ packages/topo-imagery-common/src/cli/cli_helper.py
@@ -6,9 +6,8 @@ from typing import Any, NamedTuple
 
 import shapely.geometry
 from linz_logger import get_log
-
-from topo_imagery_common.datetimes import parse_rfc_3339_date
-from topo_imagery_common.files.fs import read
+from datetimes import parse_rfc_3339_date
+from files.fs import read
 
 
 class InputParameterError(Exception):
