+++ packages/topo-imagery-common/src/cli/tests/cli_helper_test.py
@@ -4,8 +4,7 @@ from typing import Any
 from pytest import raises
 from pytest_subtests import SubTests
 from shapely.geometry import MultiPolygon
-
-from topo_imagery_common.cli.cli_helper import (
+from cli.cli_helper import (
     TileFiles,
     coalesce_multi_single,
     get_geometry_from_geojson_feature,
