+++ packages/topo-imagery-common/src/files/files_helper.py
