+++ packages/topo-imagery-common/src/files/fs_local.py
