+++ packages/topo-imagery-common/src/files/fs_s3.py
@@ -5,10 +5,9 @@ from typing import TYPE_CHECKING, Any
 
 from boto3 import client
 from linz_logger import get_log
-
-from topo_imagery_common.aws.aws_helper import get_session, parse_path
-from topo_imagery_common.files import checksum
-from topo_imagery_common.logging.time_helper import time_in_ms
+from aws.aws_helper import get_session, parse_path
+from files import checksum
+from log.time_helper import time_in_ms
 
 if TYPE_CHECKING:
     from mypy_boto3_s3 import S3Client
