+++ packages/topo-imagery-common/src/files/tests/fs_local_test.py
@@ -1,8 +1,7 @@
 import os
 
 import pytest
-
-from topo_imagery_common.files.fs_local import exists, read, write
+from files.fs_local import exists, read, write
 
 
 @pytest.mark.dependency(name="write")
