+++ scripts/generate_hillshade.py
@@ -5,18 +5,18 @@ from datetime import datetime, timezone
 from functools import partial
 from multiprocessing import Pool
 
+from cli.cli_helper import InputParameterError, TileFiles, load_input_files
+from cli.common_args import CommonArgumentParser
+from datetimes import RFC_3339_DATETIME_FORMAT
+from files.files_helper import SUFFIX_JSON, ContentType, is_tiff
+from files.fs import exists, read, write, write_all
 from linz_logger import get_log
+from log.time_helper import time_in_ms
 
-from topo_imagery_common.cli.cli_helper import InputParameterError, TileFiles, load_input_files
-from topo_imagery_common.cli.common_args import CommonArgumentParser
-from topo_imagery_common.datetimes import RFC_3339_DATETIME_FORMAT
-from topo_imagery_common.files.files_helper import SUFFIX_JSON, ContentType, is_tiff
-from topo_imagery_common.files.fs import exists, read, write, write_all
 from scripts.gdal.gdal_commands import get_gdal_command, get_hillshade_command
 from scripts.gdal.gdal_helper import run_gdal
 from scripts.gdal.gdal_presets import CompressionPreset, HillshadePreset
 from scripts.json_codec import dict_to_json_bytes
-from topo_imagery_common.logging.time_helper import time_in_ms
 from scripts.stac.imagery.create_stac import create_item
 from scripts.standardising import create_vrt
