+++ scripts/pdal/pdal_commands.py
@@ -3,11 +3,10 @@ import subprocess
 from shutil import rmtree
 from tempfile import mkdtemp
 
+from aws.aws_helper import is_s3
+from files.fs import copy
 from linz_logger import get_log
-
-from topo_imagery_common.aws.aws_helper import is_s3
-from topo_imagery_common.files.fs import copy
-from topo_imagery_common.logging.time_helper import time_in_ms
+from log.time_helper import time_in_ms
 
 
 class PDALExecutionException(Exception):
