+++ scripts/stac/imagery/collection.py
@@ -4,14 +4,15 @@ from decimal import Decimal
 from typing import Any
 
 import ulid
+from datetimes import convert_utc_to_nz_datetime, format_rfc_3339_datetime_string, parse_rfc_3339_datetime
+from files import checksum
+from files.files_helper import ContentType
+from files.fs import exists, read, write
 from linz_logger import get_log
 from shapely import to_geojson
 from shapely.geometry import shape
 from shapely.geometry.base import BaseGeometry
 
-from topo_imagery_common.datetimes import convert_utc_to_nz_datetime, format_rfc_3339_datetime_string, parse_rfc_3339_datetime
-from topo_imagery_common.files.files_helper import ContentType
-from topo_imagery_common.files.fs import exists, read, write
 from scripts.json_codec import dict_to_json_bytes
 from scripts.stac.imagery.capture_area import generate_capture_area
 from scripts.stac.imagery.collection_context import CollectionContext
@@ -36,7 +37,6 @@ from scripts.stac.imagery.constants import (
 )
 from scripts.stac.imagery.provider import Provider
 from scripts.stac.link import Link, Relation
-from topo_imagery_common.files import checksum
 from scripts.stac.util.STAC_VERSION import STAC_VERSION
 from scripts.stac.util.media_type import StacMediaType
 from scripts.stac.util.stac_extensions import StacExtensions
