+++ scripts/stac/imagery/tests/collection_test.py
@@ -9,6 +9,9 @@ from tempfile import mkdtemp
 import pytest
 import shapely.geometry
 from boto3 import client
+from files.files_helper import ContentType
+from files.fs import read
+from files.fs_s3 import write
 from moto import mock_aws
 from moto.s3.responses import DEFAULT_REGION_NAME
 from mypy_boto3_s3 import S3Client
@@ -16,9 +19,6 @@ from pytest import CaptureFixture, mark, param
 from pytest_subtests import SubTests
 
 from scripts.conftest import any_epoch_datetime_string, fake_linz_slug
-from topo_imagery_common.files.files_helper import ContentType
-from topo_imagery_common.files.fs import read
-from topo_imagery_common.files.fs_s3 import write
 from scripts.stac.imagery.collection import WARN_NO_PUBLISHED_CAPTURE_AREA, ImageryCollection, MissingMetadataError
 from scripts.stac.imagery.collection_context import CollectionContext
 from scripts.stac.imagery.item import ImageryItem, STACAsset
