+++ scripts/stac/imagery/tests/create_stac_test.py
@@ -3,10 +3,10 @@ from datetime import timedelta
 from pathlib import Path
 from typing import Any, cast
 
+from datetimes import format_rfc_3339_datetime_string
 from pytest_subtests import SubTests
 
 from scripts.conftest import any_epoch_datetime, any_epoch_datetime_string
-from topo_imagery_common.datetimes import format_rfc_3339_datetime_string
 from scripts.gdal.gdalinfo import GdalInfo
 from scripts.stac.imagery.collection import ImageryCollection
 from scripts.stac.imagery.collection_context import CollectionContext
