+++ scripts/stac/imagery/tests/item_test.py
@@ -4,10 +4,10 @@ from pathlib import Path
 from typing import Any
 from unittest.mock import patch
 
+from files.files_helper import get_file_name_from_path
 from pytest_subtests import SubTests
 
 from scripts.conftest import any_epoch_datetime_string
-from topo_imagery_common.files.files_helper import get_file_name_from_path
 from scripts.stac.imagery.collection import ImageryCollection
 from scripts.stac.imagery.collection_context import CollectionContext
 from scripts.stac.imagery.item import ImageryItem
