+++ scripts/stac/link.py
@@ -1,6 +1,7 @@
 from enum import Enum
 
-from topo_imagery_common.files import checksum
+from files import checksum
+
 from scripts.stac.util.media_type import StacMediaType
