+++ scripts/standardise_validate.py
@@ -3,9 +3,7 @@ import sys
 from argparse import ArgumentParser
 from datetime import datetime, timezone
 
-from linz_logger import get_log
-
-from topo_imagery_common.cli.cli_helper import (
+from cli.cli_helper import (
     InputParameterError,
     load_input_files,
     str_to_bool,
@@ -14,14 +12,16 @@ from topo_imagery_common.cli.cli_helper import (
     str_to_positive_int,
     valid_date,
 )
-from topo_imagery_common.datetimes import RFC_3339_DATETIME_FORMAT, format_rfc_3339_nz_midnight_datetime_string
-from scripts.tiff.file_tiff import FileTiff
-from topo_imagery_common.files.files_helper import SUFFIX_JSON, ContentType
-from topo_imagery_common.files.fs import exists, write
+from datetimes import RFC_3339_DATETIME_FORMAT, format_rfc_3339_nz_midnight_datetime_string
+from files.files_helper import SUFFIX_JSON, ContentType
+from files.fs import exists, write
+from linz_logger import get_log
+
 from scripts.gdal.gdal_helper import get_srs, get_vfs_path
 from scripts.json_codec import dict_to_json_bytes
 from scripts.stac.imagery.create_stac import create_item
 from scripts.standardising import StandardisingConfig, run_standardising
+from scripts.tiff.file_tiff import FileTiff
 
 
 def get_args_parser() -> ArgumentParser:
